I promise, no more changes V1

+++ src/use cases/remove-from-wishlist.js
@@ -2,15 +2,15 @@
 const userDbService = require("../services/users-db-service");
 
 module.exports = async (currentUserId, itemId) => {
-  //check item in whislist
-
+  // check item in wishlist
   const wishList = await userDbService.getWishList(currentUserId);
 
-  // whislist is an array , i need to loop to check of itemid.
+  // check if itemId exists in wishList
+  const itemExists = wishList.some((item) => item.item_id === itemId);
 
-  console.log(wishList);
-
-  // remove item to whislist
+  if (!itemExists) {
+    throw new Error("Item not found in wishlist");
+  }
 
   await userDbService.removeItemWhised(itemId);
 };
